refactor: extract the value-viewer subsystem from @endo/chat into @endo/spaces-util
First slice of moving @endo/chat's view components into @endo/spaces-util (the
shared chat-UI package), leaving chat as dispatch/mount + gutter + new-space
modal. The value-viewer takes `powers` as a prop and imports `E()` /
`renderConfined` directly — no chat-only coupling.

Moved (git mv) into packages/spaces-util/src/: value-component.js,
value-render.js, language-detect.js, markdown-preview.js. (blob-viewer stays in
chat — it's composer chrome.)

- Repointed the confine helpers off chat's setup-preact-container bundle to the
  real sources: h ← preact, useState ← preact/hooks,
  renderConfined/unmount ← @endo/preact-container/renderer; self
  @endo/spaces-util/* imports → relative.
- spaces-util deps += @endo/monaco-wrapper, marked (only what the moved files
  import; the rest were already present); exports + index.js re-export
  valueComponent.
- chat.js imports valueComponent from @endo/spaces-util; blob-viewer repoints
  language-detect/markdown-preview to @endo/spaces-util. Tests repointed in
  place (spaces-util has no test harness yet).

Self-contained closure (no unexpected chat module dragged in). vite build
resolves the new imports; spaces-util + chat lint 0; 74 value-viewer/blob tests
pass.
