+++ packages/chat/blob-viewer.js
@@ -8,8 +8,8 @@
 import { E } from '@endo/far';
 import { createMonacoEditor, detectTheme } from '@endo/monaco-wrapper';
 import { MarkdownFragment } from '@endo/spaces-util/markdown-vnodes.js';
-import { inferLanguage } from './language-detect.js';
-import { isMarkdown } from './markdown-preview.js';
+import { inferLanguage } from '@endo/spaces-util/language-detect.js';
+import { isMarkdown } from '@endo/spaces-util/markdown-preview.js';
 import { keyCombo, modKey } from './platform-keys.js';
 
 import { h, renderConfined, unmount } from './setup-preact-container.js';
