+++ packages/spaces-util/src/language-detect.js
