+++ packages/spaces-util/src/value-render.js
@@ -1,7 +1,7 @@
 // @ts-check
 
 import { passStyleOf, getInterfaceOf } from '@endo/pass-style';
-import { numberFormatter } from '@endo/spaces-util/time-formatters.js';
+import { numberFormatter } from './time-formatters.js';
 
 /**
  * Render a value as DOM elements.
